Setting up repo
Created the docs folder that will host the website html and put a sample index.html file in there based on covid-19.trialstracker.net that can be edited and built off of.

Also created a data folder that will hold the CSV and JSON files.

I've also included a basic python script. Once you've cloned the repo locally, you should be able to put the CSV in the data folder, put the path where I've indicated in the .py file, and then run that to produce the JSON.